remove unnecessary correct_rightmost_block parameter and related tests
